+++ tests/Infrastructure/EpicDiscovery.Tests/EpicMetadataFetcherTests.cs
@@ -1,3 +1,4 @@
+using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Text;
@@ -42,23 +43,47 @@ public sealed class EpicMetadataFetcherTests
 
         var handler = new StubHttpMessageHandler((request, _) =>
         {
-            request.Method.Should().Be(HttpMethod.Post);
-            request.RequestUri!.Should().Be(new Uri("https://graphql.epicgames.com/graphql"));
-            var response = new HttpResponseMessage(HttpStatusCode.OK)
+            if (request.Method == HttpMethod.Post)
             {
-                Content = new StringContent(responseJson, Encoding.UTF8, "application/json"),
+                request.RequestUri!.Should().Be(new Uri("https://graphql.epicgames.com/graphql"));
+                var response = new HttpResponseMessage(HttpStatusCode.OK)
+                {
+                    Content = new StringContent(responseJson, Encoding.UTF8, "application/json"),
+                };
+                return Task.FromResult(response);
+            }
+
+            request.RequestUri!.ToString().Should().Be("https://cdn.epicgames.com/fn/diesel.jpg");
+            var imageResponse = new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new ByteArrayContent(new byte[] { 1, 2, 3 })
             };
-            return Task.FromResult(response);
+            return Task.FromResult(imageResponse);
         });
 
-        var fetcher = new EpicMetadataFetcher(new HttpClient(handler));
+        var cacheDirectory = CreateTemporaryCacheDirectory();
+        try
+        {
+            var httpClient = new HttpClient(handler);
+            var heroArtCache = new EpicHeroArtCache(httpClient, new TestFileAccessor(), cacheDirectory: cacheDirectory);
+            var fetcher = new EpicMetadataFetcher(httpClient, heroArtCache);
 
-        var catalogItem = await fetcher.FetchAsync(entitlement);
+            var catalogItem = await fetcher.FetchAsync(entitlement);
 
-        catalogItem.Should().NotBeNull();
-        catalogItem!.Title.Should().Be("Fortnite Deluxe");
-        catalogItem.ProductSlug.Should().Be("fortnite");
-        catalogItem.KeyImages.Should().ContainSingle(image => image.Type == "DieselGameBox" && image.Uri!.EndsWith("diesel.jpg"));
+            catalogItem.Should().NotBeNull();
+            catalogItem!.Title.Should().Be("Fortnite Deluxe");
+            catalogItem.ProductSlug.Should().Be("fortnite");
+            var image = catalogItem.KeyImages.Should().ContainSingle()
+                .Which;
+            image.Type.Should().Be("DieselGameBox");
+            image.Uri.Should().EndWith("diesel.jpg");
+            image.Path.Should().NotBeNullOrWhiteSpace();
+            File.Exists(image.Path!).Should().BeTrue();
+        }
+        finally
+        {
+            CleanupDirectory(cacheDirectory);
+        }
     }
 
     [Fact]
@@ -81,26 +106,69 @@ public sealed class EpicMetadataFetcherTests
                 return Task.FromResult(failure);
             }
 
-            request.Method.Should().Be(HttpMethod.Get);
-            request.RequestUri!.ToString().Should().Be("https://store-content.ak.epicgames.com/api/en-US/content/products/rocket-league");
-            var contentResponse = new HttpResponseMessage(HttpStatusCode.OK)
+            if (request.RequestUri!.ToString().Equals("https://store-content.ak.epicgames.com/api/en-US/content/products/rocket-league"))
             {
-                Content = new StringContent(
-                    "{ \"title\": \"Rocket League Content\", \"keyImages\": [{ \"type\": \"OfferImageWide\", \"url\": \"https://cdn.epicgames.com/rocket/wide.jpg\" }] }",
-                    Encoding.UTF8,
-                    "application/json"),
+                var contentResponse = new HttpResponseMessage(HttpStatusCode.OK)
+                {
+                    Content = new StringContent(
+                        "{ \"title\": \"Rocket League Content\", \"keyImages\": [{ \"type\": \"OfferImageWide\", \"url\": \"https://cdn.epicgames.com/rocket/wide.jpg\" }] }",
+                        Encoding.UTF8,
+                        "application/json"),
+                };
+                return Task.FromResult(contentResponse);
+            }
+
+            request.RequestUri!.ToString().Should().Be("https://cdn.epicgames.com/rocket/wide.jpg");
+            var imageResponse = new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new ByteArrayContent(new byte[] { 4, 5, 6 })
             };
-            return Task.FromResult(contentResponse);
+            return Task.FromResult(imageResponse);
         });
 
-        var fetcher = new EpicMetadataFetcher(new HttpClient(handler));
+        var cacheDirectory = CreateTemporaryCacheDirectory();
+        try
+        {
+            var httpClient = new HttpClient(handler);
+            var heroArtCache = new EpicHeroArtCache(httpClient, new TestFileAccessor(), cacheDirectory: cacheDirectory);
+            var fetcher = new EpicMetadataFetcher(httpClient, heroArtCache);
 
-        var catalogItem = await fetcher.FetchAsync(entitlement);
+            var catalogItem = await fetcher.FetchAsync(entitlement);
 
-        catalogItem.Should().NotBeNull();
-        catalogItem!.Title.Should().Be("Rocket League Content");
-        catalogItem.ProductSlug.Should().Be("rocket-league");
-        catalogItem.KeyImages.Should().ContainSingle(image => image.Type == "OfferImageWide" && image.Uri == "https://cdn.epicgames.com/rocket/wide.jpg");
-        handler.Requests.Should().HaveCount(2);
+            catalogItem.Should().NotBeNull();
+            catalogItem!.Title.Should().Be("Rocket League Content");
+            catalogItem.ProductSlug.Should().Be("rocket-league");
+            var image = catalogItem.KeyImages.Should().ContainSingle()
+                .Which;
+            image.Type.Should().Be("OfferImageWide");
+            image.Path.Should().NotBeNullOrWhiteSpace();
+            handler.Requests.Should().HaveCount(3);
+        }
+        finally
+        {
+            CleanupDirectory(cacheDirectory);
+        }
+    }
+
+    private static string CreateTemporaryCacheDirectory()
+    {
+        var path = Path.Combine(Path.GetTempPath(), "EpicMetadataFetcherTests", Guid.NewGuid().ToString("N"));
+        Directory.CreateDirectory(path);
+        return path;
+    }
+
+    private static void CleanupDirectory(string path)
+    {
+        try
+        {
+            if (Directory.Exists(path))
+            {
+                Directory.Delete(path, recursive: true);
+            }
+        }
+        catch
+        {
+            // ignore cleanup failures
+        }
     }
 }